App reordering

+++ lib/ui/home/tabs/test_tab.dart
@@ -47,7 +47,9 @@ class TestTab extends HookWidget implements CobbleScreen {
     );
 
     final appManager = useProvider(appManagerProvider);
-    final allApps = useProvider(appManagerProvider.state);
+    final allPackages = useProvider(appManagerProvider.state);
+    final allApps =
+        allPackages.where((element) => !element.isWatchface).toList();
 
     useEffect(() {
       Future.microtask(() async {
@@ -190,7 +192,7 @@ class TestTab extends HookWidget implements CobbleScreen {
                   [],
               Text("Disable BLE Workarounds: "),
               ...neededWorkarounds.map(
-                    (workaround) => Row(children: [
+                (workaround) => Row(children: [
                   Switch(
                     value: workaround.disabled,
                     onChanged: (value) async {
@@ -201,11 +203,49 @@ class TestTab extends HookWidget implements CobbleScreen {
                   Text(workaround.name)
                 ]),
               ),
+              Text("Installed watchfaces: "),
+              ...allPackages.where((element) => element.isWatchface).map(
+                (face) {
+                  String compatibleText = "";
+                  final currentWatch = connectionState.currentConnectedWatch;
+
+                  if (currentWatch != null) {
+                    final watchType = currentWatch
+                        .runningFirmware.hardwarePlatform
+                        .getWatchType();
+
+                    if (face.isCompatibleWith(watchType)) {
+                      compatibleText = " (Compatible)";
+                    } else {
+                      compatibleText = " (Incompatible)";
+                    }
+                  }
+
+                  return Row(children: [
+                    Container(
+                      margin: EdgeInsets.all(16),
+                      child: Text(
+                          "${face.longName}$compatibleText by ${face.company}"),
+                    ),
+                    if (!face.isSystem)
+                      ElevatedButton(
+                        child: Text("Delete"),
+                        onPressed: () {
+                          appManager.deleteApp(face.uuid);
+                        },
+                      ),
+                  ]);
+                },
+              ),
               Text("Installed apps: "),
               ...allApps.map(
                 (app) {
                   String compatibleText = "";
                   final currentWatch = connectionState.currentConnectedWatch;
+
+                  final atTop = app.appOrder == 0;
+                  final atBottom = app.appOrder == allApps.length - 1;
+
                   if (currentWatch != null) {
                     final watchType = currentWatch
                         .runningFirmware.hardwarePlatform
@@ -224,12 +264,33 @@ class TestTab extends HookWidget implements CobbleScreen {
                       child: Text(
                           "${app.longName}$compatibleText by ${app.company}"),
                     ),
-                    ElevatedButton(
-                      child: Text("Delete"),
-                      onPressed: () {
-                        appManager.deleteApp(app.uuid);
-                      },
-                    )
+                    if (!app.isSystem)
+                      ElevatedButton(
+                        child: Text("Delete"),
+                        onPressed: () {
+                          appManager.deleteApp(app.uuid);
+                        },
+                      ),
+                    if (!atTop)
+                      Container(
+                        margin: EdgeInsets.only(left: 8),
+                        child: ElevatedButton(
+                          child: Text("Up"),
+                          onPressed: () {
+                            appManager.reorderApp(app.uuid, app.appOrder - 1);
+                          },
+                        ),
+                      ),
+                    if (!atBottom)
+                      Container(
+                        margin: EdgeInsets.only(left: 8),
+                        child: ElevatedButton(
+                          child: Text("Down"),
+                          onPressed: () {
+                            appManager.reorderApp(app.uuid, app.appOrder + 1);
+                          },
+                        ),
+                      ),
                   ]);
                 },
               )
